Refactor rule schema structure
